refactor(translation): extract TranslationPromptBuilder, share translation preamble
Split prompt construction (buildTranslationPrompt + getLanguageName) into
a dedicated TranslationPromptBuilder collaborator, and fold the shared
guard/validate/detect/build preamble of translate() and
translateForConfiguration() into one prepareTranslation() helper. This
removes the duplication between the two public methods and keeps
TranslationService within its method-count budget. The public
TranslationServiceInterface is unchanged.
