+++ Tests/Unit/Service/Feature/TranslationServiceTest.php
@@ -18,6 +18,7 @@ use Netresearch\NrLlm\Exception\ConfigurationNotFoundException;
 use Netresearch\NrLlm\Exception\InvalidArgumentException;
 use Netresearch\NrLlm\Provider\Middleware\BudgetMiddleware;
 use Netresearch\NrLlm\Service\Budget\BackendUserContextResolverInterface;
+use Netresearch\NrLlm\Service\Feature\TranslationPromptBuilder;
 use Netresearch\NrLlm\Service\Feature\TranslationService;
 use Netresearch\NrLlm\Service\LlmConfigurationServiceInterface;
 use Netresearch\NrLlm\Service\LlmServiceManagerInterface;
@@ -55,6 +56,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $this->llmManagerStub,
             $this->translatorRegistryMock,
             $this->configServiceStub,
+            new TranslationPromptBuilder(),
         );
     }
 
@@ -128,6 +130,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $llmManagerMock,
             $this->translatorRegistryMock,
             $this->configServiceStub,
+            new TranslationPromptBuilder(),
         );
 
         $result = $subject->translate('Hello', 'de');
@@ -299,6 +302,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $llmManagerMock,
             $this->translatorRegistryMock,
             $this->configServiceStub,
+            new TranslationPromptBuilder(),
         );
 
         $result = $subject->translateBatch(['One', 'Two', 'Three'], 'de', 'en');
@@ -652,6 +656,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $llmManagerMock,
             $this->translatorRegistryMock,
             $this->configServiceStub,
+            new TranslationPromptBuilder(),
         );
 
         $options = new TranslationOptions(formality: 'formal');
@@ -879,6 +884,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $this->llmManagerStub,
             $translatorRegistryMock,
             $configServiceMock,
+            new TranslationPromptBuilder(),
         );
 
         $reflection = new ReflectionClass($subject);
@@ -910,6 +916,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $this->llmManagerStub,
             $translatorRegistryMock,
             $configServiceMock,
+            new TranslationPromptBuilder(),
         );
 
         $reflection = new ReflectionClass($subject);
@@ -945,6 +952,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $this->llmManagerStub,
             $translatorRegistryMock,
             $configServiceMock,
+            new TranslationPromptBuilder(),
         );
 
         $reflection = new ReflectionClass($subject);
@@ -986,6 +994,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $llmManagerMock,
             $this->translatorRegistryMock,
             $this->configServiceStub,
+            new TranslationPromptBuilder(),
         );
 
         $options = new TranslationOptions(formality: 'informal');
@@ -1072,6 +1081,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $llmManagerMock,
             $this->translatorRegistryMock,
             $this->configServiceStub,
+            new TranslationPromptBuilder(),
             $resolver,
         );
 
@@ -1100,6 +1110,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $llmManagerMock,
             $this->translatorRegistryMock,
             $this->configServiceStub,
+            new TranslationPromptBuilder(),
             $resolver,
         );
 
@@ -1133,6 +1144,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $llmManagerMock,
             $this->translatorRegistryMock,
             $this->configServiceStub,
+            new TranslationPromptBuilder(),
             $resolver,
         );
 
@@ -1161,6 +1173,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $llmManagerMock,
             $this->translatorRegistryMock,
             $this->configServiceStub,
+            new TranslationPromptBuilder(),
             $resolver,
         );
 
@@ -1223,6 +1236,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $this->llmManagerStub,
             $this->translatorRegistryMock,
             $this->configServiceStub,
+            new TranslationPromptBuilder(),
             $resolver,
         );
 
@@ -1345,6 +1359,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $llmManagerMock,
             $this->translatorRegistryMock,
             $this->configServiceStub,
+            new TranslationPromptBuilder(),
         );
 
         $call($subject);
@@ -1352,29 +1367,6 @@ class TranslationServiceTest extends AbstractUnitTestCase
         self::assertInstanceOf(ChatOptions::class, $capturedOptions);
 
         return ['messages' => $capturedMessages, 'options' => $capturedOptions];
-    }
-
-    /**
-     * Invoke the private buildTranslationPrompt() with a fully-controlled raw
-     * options array and return the ['system', 'user'] strings.
-     *
-     * @param array<string, mixed> $options
-     *
-     * @return array{system: string, user: string}
-     */
-    private function invokeBuildPrompt(string $text, string $source, string $target, array $options): array
-    {
-        $reflection = new ReflectionClass($this->subject);
-        $method = $reflection->getMethod('buildTranslationPrompt');
-        $prompt = $method->invoke($this->subject, $text, $source, $target, $options);
-
-        self::assertIsArray($prompt);
-        $system = $prompt['system'] ?? null;
-        $user = $prompt['user'] ?? null;
-        self::assertIsString($system);
-        self::assertIsString($user);
-
-        return ['system' => $system, 'user' => $user];
     }
 
     #[Test]
@@ -1543,6 +1535,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $this->llmManagerStub,
             $translatorRegistryMock,
             $configServiceMock,
+            new TranslationPromptBuilder(),
         );
 
         $reflection = new ReflectionClass($subject);
@@ -1550,118 +1543,6 @@ class TranslationServiceTest extends AbstractUnitTestCase
         $result = $method->invoke($subject, ['configuration' => '']);
 
         self::assertSame($translatorStub, $result);
-    }
-
-    #[Test]
-    public function buildTranslationPromptRendersMinimalPrompt(): void
-    {
-        $prompt = $this->invokeBuildPrompt('Hello World', 'en', 'de', [
-            'formality' => 'default',
-            'domain' => 'general',
-            'glossary' => [],
-            'context' => '',
-            'preserve_formatting' => true,
-        ]);
-
-        $system = $prompt['system'];
-        // Domain + language-name resolution (en -> English, de -> German).
-        self::assertStringContainsString(
-            'You are a professional general translator. Translate the following text from English to German.',
-            $system,
-        );
-        // formality 'default' => no tone line.
-        self::assertStringNotContainsString('Maintain', $system);
-        // preserve_formatting true => formatting line present.
-        self::assertStringContainsString(
-            'Preserve all formatting, HTML tags, markdown, and special characters.',
-            $system,
-        );
-        // Empty glossary / context => their sections are absent.
-        self::assertStringNotContainsString('Use these exact term translations:', $system);
-        self::assertStringNotContainsString('Context (for reference only):', $system);
-        // Trailing instruction is appended, not replacing the prompt.
-        self::assertStringContainsString('Provide ONLY the translation, no explanations or notes.', $system);
-
-        self::assertSame("Translate this text:\n\nHello World", $prompt['user']);
-    }
-
-    #[Test]
-    public function buildTranslationPromptDefaultsPreserveFormattingToTrueWhenKeyAbsent(): void
-    {
-        // No 'preserve_formatting' key => the `?? true` default applies.
-        $prompt = $this->invokeBuildPrompt('Hello', 'en', 'de', [
-            'formality' => 'default',
-            'domain' => 'general',
-        ]);
-
-        self::assertStringContainsString(
-            'Preserve all formatting, HTML tags, markdown, and special characters.',
-            $prompt['system'],
-        );
-    }
-
-    #[Test]
-    public function buildTranslationPromptOmitsFormattingLineWhenPreserveFormattingFalse(): void
-    {
-        $prompt = $this->invokeBuildPrompt('Hello', 'en', 'de', [
-            'formality' => 'default',
-            'domain' => 'general',
-            'preserve_formatting' => false,
-        ]);
-
-        self::assertStringNotContainsString('Preserve all formatting', $prompt['system']);
-    }
-
-    #[Test]
-    public function buildTranslationPromptRendersAllSectionsWithFullOptions(): void
-    {
-        $prompt = $this->invokeBuildPrompt('Hello', 'en', 'de', [
-            'formality' => 'formal',
-            'domain' => 'legal',
-            'glossary' => [
-                'Hello' => 'Hallo',   // string value
-                'count' => 5,         // int value
-                'ratio' => 1.5,       // float value
-                'bad' => ['x'],       // non-scalar => filtered out
-            ],
-            'context' => 'Software docs',
-            'preserve_formatting' => false,
-        ]);
-
-        $system = $prompt['system'];
-
-        // Intro must survive every section append (guards against `.=` -> `=`).
-        self::assertStringContainsString(
-            'You are a professional legal translator. Translate the following text from English to German.',
-            $system,
-        );
-        // formality 'formal' => tone line.
-        self::assertStringContainsString('Maintain formal tone.', $system);
-        // preserve false => no formatting line.
-        self::assertStringNotContainsString('Preserve all formatting', $system);
-        // Glossary header + one line per scalar term.
-        self::assertStringContainsString('Use these exact term translations:', $system);
-        self::assertStringContainsString('- Hello → Hallo', $system);
-        self::assertStringContainsString('- count → 5', $system);
-        self::assertStringContainsString('- ratio → 1.5', $system);
-        // Non-scalar glossary value is filtered out entirely.
-        self::assertStringNotContainsString('- bad', $system);
-        // Context section rendered.
-        self::assertStringContainsString("Context (for reference only):\nSoftware docs", $system);
-        // Closing instruction still appended.
-        self::assertStringContainsString('Provide ONLY the translation, no explanations or notes.', $system);
-    }
-
-    #[Test]
-    public function buildTranslationPromptResolvesKnownLanguageNames(): void
-    {
-        // getLanguageName maps known codes; an unknown code falls back to itself.
-        $prompt = $this->invokeBuildPrompt('Hi', 'fr', 'zz', [
-            'formality' => 'default',
-            'domain' => 'general',
-        ]);
-
-        self::assertStringContainsString('from French to zz.', $prompt['system']);
     }
 
     // ==================== #429: configured model reaches ChatOptions ====================
@@ -1773,6 +1654,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $llmManagerMock,
             $this->translatorRegistryMock,
             $this->configServiceStub,
+            new TranslationPromptBuilder(),
             $resolver,
         );
 
@@ -1824,6 +1706,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $llmManagerMock,
             $this->translatorRegistryMock,
             $this->configServiceStub,
+            new TranslationPromptBuilder(),
         );
 
         $result = $subject->translateForConfiguration('Hello World', 'de', $configuration, 'en');
@@ -1982,6 +1865,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $llmManagerMock,
             $this->translatorRegistryMock,
             $this->configServiceStub,
+            new TranslationPromptBuilder(),
         );
 
         $result = $subject->translateForConfiguration('Bonjour', 'de', $configuration);
@@ -2051,6 +1935,7 @@ class TranslationServiceTest extends AbstractUnitTestCase
             $this->llmManagerStub,
             $translatorRegistryMock,
             $configServiceMock,
+            new TranslationPromptBuilder(),
         );
 
         $result = $subject->translateWithTranslator(
